admin/requirements/base.txt
boto3==1.26.1
coreapi==2.3.3
django==4.2.19
django-cms==3.11.0
django-cors-headers==4.7.0
django-extensions==3.2.1
django-health-check==3.17.0
djangocms-text-ckeditor==5.1.1
djangocms-file==3.0.0
django-filer==2.2.3
django-filter==22.1
django-otp==1.5.4
django-recaptcha==3.0.0
djangorestframework==3.14.0
django-rest-knox==4.2.0
django-ses==3.2.2
django-soft-delete==0.9.21
django-storages==1.13.1
git+https://github.com/sunscrapers/djoser.git@abdf622f95dfa2c6278c4bd6d50dfe69559d90c0
drf-recaptcha==3.2.0
google-auth==2.16.1
gunicorn==20.1.0
psycopg2-binary==2.9.5
redis==4.3.4